Move ValidationCase to cases folder
refs #30819
