+++ python/TestHarness/validation/__init__.py
@@ -7,4 +7,4 @@
 #* Licensed under LGPL 2.1, please see LICENSE for details
 #* https://www.gnu.org/licenses/lgpl-2.1.html
 
-from .case import ValidationCase
+from .cases.validationcase import ValidationCase
